Allow for Jira ticket in title
Since a PR is opened, it cannot have it's base branch edited. That means
that if we forget to put a PR number, we have to close it and open it
again.

With this change, it is now possible to put the ticket in the PR title
to be picked up as well and linked properly.